// packages/nx-plugin/src/cedar/agentcore-gateway/schema.d.ts
/**
 * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: Apache-2.0
 */
import type { IacOption } from '../../utils/iac';

export interface AgentcoreGatewayGeneratorSchema {
  name: string;
  directory?: string;
  subDirectory?: string;
  protocol?: 'mcp';
  auth?: 'iam';
  iac: IacOption;
}